Review e organização de codigo
