+++ src/context/ModalsContext.jsx
@@ -34,6 +34,8 @@ export const ModalsProvider = ({ children }) => {
 	const [editFinished, SetEditFinished] = useState(false);
 	const [showErrorSearch, setShowErrorSearch] = useState(false);
 	const [showErrorBilling, setShowErrorBilling] = useState(false);
+	const [showPassword, setShowPassword] = useState(false)
+	const [passowrdCombination, setPassowrdCombination] = useState("")
 
 	const handleOpenEditUser = () => setOpenModalEditUser(true);
 	const handleCloseEditUser = () => setOpenModalEditUser(false);
@@ -85,11 +87,15 @@ export const ModalsProvider = ({ children }) => {
 				openSnackChargeDelete,
 				setOpenSnackChargeDelete,
 				openChargeDetailsModal,
-				setOpenChargeDetailsModal, 
+				setOpenChargeDetailsModal,
 				showErrorSearch,
 				setShowErrorSearch,
 				showErrorBilling,
-				setShowErrorBilling
+				setShowErrorBilling,
+				showPassword,
+				setShowPassword,
+				passowrdCombination,
+				setPassowrdCombination,
 			}}
 		>
 			{children}
